refactor: ec2 SD, move SD struct from config into the ec2 package
