+++ discovery/ec2/ec2.go
@@ -32,8 +32,10 @@ import (
 	"github.com/prometheus/common/model"
 
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/prometheus/prometheus/config"
+	"github.com/prometheus/prometheus/pkg/targetgroup"
+	configUtil "github.com/prometheus/prometheus/util/config"
 	"github.com/prometheus/prometheus/util/strutil"
+	yamlUtil "github.com/prometheus/prometheus/util/yaml"
 )
 
 const (
@@ -62,7 +64,52 @@ var (
 			Name: "prometheus_sd_ec2_refresh_duration_seconds",
 			Help: "The duration of a EC2-SD refresh in seconds.",
 		})
+	// DefaultSDConfig is the default EC2 SD configuration.
+	DefaultSDConfig = SDConfig{
+		Port:            80,
+		RefreshInterval: model.Duration(60 * time.Second),
+	}
 )
+
+// SDConfig is the configuration for EC2 based service discovery.
+type SDConfig struct {
+	Region          string            `yaml:"region"`
+	AccessKey       string            `yaml:"access_key,omitempty"`
+	SecretKey       configUtil.Secret `yaml:"secret_key,omitempty"`
+	Profile         string            `yaml:"profile,omitempty"`
+	RoleARN         string            `yaml:"role_arn,omitempty"`
+	RefreshInterval model.Duration    `yaml:"refresh_interval,omitempty"`
+	Port            int               `yaml:"port"`
+
+	// Catches all undefined fields and must be empty after parsing.
+	XXX map[string]interface{} `yaml:",inline"`
+}
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface.
+func (c *SDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	*c = DefaultSDConfig
+	type plain SDConfig
+	err := unmarshal((*plain)(c))
+	if err != nil {
+		return err
+	}
+	if err := yamlUtil.CheckOverflow(c.XXX, "ec2_sd_config"); err != nil {
+		return err
+	}
+	if c.Region == "" {
+		sess, err := session.NewSession()
+		if err != nil {
+			return err
+		}
+		metadata := ec2metadata.New(sess)
+		region, err := metadata.Region()
+		if err != nil {
+			return fmt.Errorf("EC2 SD configuration requires a region")
+		}
+		c.Region = region
+	}
+	return nil
+}
 
 func init() {
 	prometheus.MustRegister(ec2SDRefreshFailuresCount)
@@ -81,7 +128,7 @@ type Discovery struct {
 }
 
 // NewDiscovery returns a new EC2Discovery which periodically refreshes its targets.
-func NewDiscovery(conf *config.EC2SDConfig, logger log.Logger) *Discovery {
+func NewDiscovery(conf *SDConfig, logger log.Logger) *Discovery {
 	creds := credentials.NewStaticCredentials(conf.AccessKey, string(conf.SecretKey), "")
 	if conf.AccessKey == "" && conf.SecretKey == "" {
 		creds = nil
@@ -103,7 +150,7 @@ func NewDiscovery(conf *config.EC2SDConfig, logger log.Logger) *Discovery {
 }
 
 // Run implements the TargetProvider interface.
-func (d *Discovery) Run(ctx context.Context, ch chan<- []*config.TargetGroup) {
+func (d *Discovery) Run(ctx context.Context, ch chan<- []*targetgroup.Group) {
 	ticker := time.NewTicker(d.interval)
 	defer ticker.Stop()
 
@@ -113,7 +160,7 @@ func (d *Discovery) Run(ctx context.Context, ch chan<- []*config.TargetGroup) {
 		level.Error(d.logger).Log("msg", "Refresh failed", "err", err)
 	} else {
 		select {
-		case ch <- []*config.TargetGroup{tg}:
+		case ch <- []*targetgroup.Group{tg}:
 		case <-ctx.Done():
 			return
 		}
@@ -129,7 +176,7 @@ func (d *Discovery) Run(ctx context.Context, ch chan<- []*config.TargetGroup) {
 			}
 
 			select {
-			case ch <- []*config.TargetGroup{tg}:
+			case ch <- []*targetgroup.Group{tg}:
 			case <-ctx.Done():
 				return
 			}
@@ -149,7 +196,7 @@ func (d *Discovery) ec2MetadataAvailable(sess *session.Session) (isAvailable boo
 	return isAvailable
 }
 
-func (d *Discovery) refresh() (tg *config.TargetGroup, err error) {
+func (d *Discovery) refresh() (tg *targetgroup.Group, err error) {
 	t0 := time.Now()
 	defer func() {
 		ec2SDRefreshDuration.Observe(time.Since(t0).Seconds())
@@ -178,7 +225,7 @@ func (d *Discovery) refresh() (tg *config.TargetGroup, err error) {
 			ec2s = ec2.New(sess)
 		}
 	}
-	tg = &config.TargetGroup{
+	tg = &targetgroup.Group{
 		Source: *d.aws.Region,
 	}
 	if err = ec2s.DescribeInstancesPages(nil, func(p *ec2.DescribeInstancesOutput, lastPage bool) bool {
